+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   * Support `emails_enabled` parameters when creating or editing groups and projects.
   * Support filtering groups based on the repository storage.
   * Support enforcing expiration on service access tokens for group hooks.
+  * Support filtering group and project packages by version.
 
 ## Deprecations